fix: resolve build issue
