+++ src/pages/index.tsx
@@ -1,9 +1,12 @@
 import type { NextPage } from "next";
 import Insurance from "./insurance";
+import NoSSRWrapper from "../Components/NoSSRWrapper";
 
 const Home: NextPage = () => {
     return (
-        <Insurance />
+        <NoSSRWrapper>
+            <Insurance />
+        </NoSSRWrapper>
     );
 };
