+++ src/pages/insurance.tsx
@@ -7,15 +7,16 @@ import styles from "../styles/insurance.module.css";
 import { insurancePageTabs as tabs } from "../constants/routes";
 import { Fragment } from "react";
 import dynamic from "next/dynamic";
+import NoSSRWrapper from "../Components/NoSSRWrapper";
 
 const InsurancePackagePlans = dynamic(
-    () => import("../Components/InsurancePackagePlans"),
+    () => import("../Components/InsurancePackagePlans/index"),
     {
         ssr: false,
     }
 );
 
-const MyPackages = dynamic(() => import("../Components/MyPackages"), {
+const MyPackages = dynamic(() => import("../Components/MyPackages/index"), {
     ssr: false,
 });
 
@@ -24,77 +25,79 @@ const Insurance: NextPage = () => {
     const tab = router.query.tab;
 
     return (
-        <Fragment>
-            <div className={styles.container}>
-                <Head>
-                    <title>Rance Protocol - Insurance</title>
-                </Head>
-                <main className={styles.main}>
-                    <div className={styles.banner}>
-                        <div className={styles.banner__container}>
-                            <h1 className={styles.banner__header}>
-                                Insurance Info
-                            </h1>
-                            <p className={styles.banner__text}>
-                                Stabilise coins at your desired price with up to
-                                2 years Insurance protection.
-                            </p>
+        <NoSSRWrapper>
+            <Fragment>
+                <div className={styles.container}>
+                    <Head>
+                        <title>Rance Protocol - Insurance</title>
+                    </Head>
+                    <main className={styles.main}>
+                        <div className={styles.banner}>
+                            <div className={styles.banner__container}>
+                                <h1 className={styles.banner__header}>
+                                    Insurance Info
+                                </h1>
+                                <p className={styles.banner__text}>
+                                    Stabilise coins at your desired price with
+                                    up to 2 years Insurance protection.
+                                </p>
+                            </div>
+                            <div className={styles.banner__image__container}>
+                                <Image
+                                    src="/insurance-banner-image.png"
+                                    alt="staking page banner"
+                                    layout="fill"
+                                    className={styles.banner__image}
+                                />
+                            </div>
                         </div>
-                        <div className={styles.banner__image__container}>
-                            <Image
-                                src="/insurance-banner-image.png"
-                                alt="staking page banner"
-                                layout="fill"
-                                className={styles.banner__image}
-                            />
+                        <div className={styles.tab}>
+                            <ul className={styles.tabs__container}>
+                                <li>
+                                    <button
+                                        onClick={() =>
+                                            router.push(
+                                                `${router.pathname}?tab=${tabs.insurancePackages}`
+                                            )
+                                        }
+                                        className={clsx({
+                                            [styles.active_tab]:
+                                                String(tab).toLowerCase() !==
+                                                tabs.myPackages,
+                                            [styles.tab_btn]: true,
+                                        })}
+                                    >
+                                        Insurance Packages
+                                    </button>
+                                </li>
+                                <li>
+                                    <button
+                                        onClick={() =>
+                                            router.push(
+                                                `${router.pathname}?tab=${tabs.myPackages}`
+                                            )
+                                        }
+                                        className={clsx({
+                                            [styles.active_tab]:
+                                                String(tab).toLowerCase() ===
+                                                tabs.myPackages,
+                                            [styles.tab_btn]: true,
+                                        })}
+                                    >
+                                        My Packages
+                                    </button>
+                                </li>
+                            </ul>
                         </div>
-                    </div>
-                    <div className={styles.tab}>
-                        <ul className={styles.tabs__container}>
-                            <li>
-                                <button
-                                    onClick={() =>
-                                        router.push(
-                                            `${router.pathname}?tab=${tabs.insurancePackages}`
-                                        )
-                                    }
-                                    className={clsx({
-                                        [styles.active_tab]:
-                                            String(tab).toLowerCase() !==
-                                            tabs.myPackages,
-                                        [styles.tab_btn]: true,
-                                    })}
-                                >
-                                    Insurance Packages
-                                </button>
-                            </li>
-                            <li>
-                                <button
-                                    onClick={() =>
-                                        router.push(
-                                            `${router.pathname}?tab=${tabs.myPackages}`
-                                        )
-                                    }
-                                    className={clsx({
-                                        [styles.active_tab]:
-                                            String(tab).toLowerCase() ===
-                                            tabs.myPackages,
-                                        [styles.tab_btn]: true,
-                                    })}
-                                >
-                                    My Packages
-                                </button>
-                            </li>
-                        </ul>
-                    </div>
-                    {String(tab).toLowerCase() !== tabs.myPackages ? (
-                        <InsurancePackagePlans />
-                    ) : (
-                        <MyPackages />
-                    )}
-                </main>
-            </div>
-        </Fragment>
+                        {String(tab).toLowerCase() !== tabs.myPackages ? (
+                            <InsurancePackagePlans />
+                        ) : (
+                            <MyPackages />
+                        )}
+                    </main>
+                </div>
+            </Fragment>
+        </NoSSRWrapper>
     );
 };
