+++ src/pages/referral.tsx
@@ -1,12 +1,17 @@
 import type { NextPage } from "next";
 import dynamic from "next/dynamic";
+import NoSSRWrapper from "../Components/NoSSRWrapper";
 
-const ReferralView = dynamic(() => import("../views/referral"), {
+const ReferralView = dynamic(() => import("../view/referral"), {
     ssr: false,
 });
 
 const Referral: NextPage = () => {
-    return <ReferralView />;
+    return (
+        <NoSSRWrapper>
+            <ReferralView />
+        </NoSSRWrapper>
+    );
 };
 
 export default Referral;
